[Form] Move  classes to Twig bridge
